+++ homepage/example.js
@@ -205,7 +205,7 @@ function trackViewingDistance(e) {
       showVideo: false,
       nearPoint: true,
       showNearPoint: true,
-      desiredDistanceCm: 40,
+      desiredDistanceCm: 60,
       desiredDistanceMonitor: true,
     },
     distanceData => {
@@ -221,7 +221,9 @@ function trackViewingDistance(e) {
           data.method
         } method. The near point is at [${data.value.nearPointCm.x} cm, ${
           data.value.nearPointCm.y
-        } cm] compared to the center of the screen.`
+        } cm] compared to the center of the screen. Latency is ${
+          data.value.latencyMs
+        } ms.`
       )
     }
   )
@@ -319,7 +321,9 @@ function trackGaze(e) {
       gazeP.innerHTML = gotData(
         `The gaze position is [${data.value.x} px, ${
           data.value.y
-        } px] at ${parseTimestamp(data.timestamp)}.`
+        } px] at ${parseTimestamp(data.timestamp)}. Latency is ${
+          data.value.latencyMs
+        } ms.`
       )
     }
   )